toolchain/build: Build both arm-eabi- and arm-linux-androideabi- compilers
This makes the toolchain build process create both an arm-eabi- and an
arm-linux-androideabi- compiler (and binutils), enabling us to use
arm-linux-androideabi- where we can while staying with arm-eabi- for
other bits.

Since u-boot should always be built with a bare metal compiler, this
is the right thing to do even once the arm-linux-androideabi toolchain
can compile the entire OS correctly.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>